Introduce transaction manager
This extracts the logic of transaction management from the safenet
onchain protocol class into a separate reusable class. This way this
logic can be reused for the sentinel service to monitor  and retry
transactions.
