+++ validator/src/__tests__/data/protocol.ts
@@ -223,7 +223,7 @@ export const TEST_ACTIONS: [ProtocolAction, keyof TestProtocol, EthTransactionDa
 		{
 			to: TEST_COORDINATOR,
 			value: 0n,
-			data: "0x0b2b35375afe0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001",
+			data: "0x2f559b6d5afe0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001",
 			gas: 300_000n,
 		},
 	],
@@ -238,7 +238,7 @@ export const TEST_ACTIONS: [ProtocolAction, keyof TestProtocol, EthTransactionDa
 		{
 			to: TEST_COORDINATOR,
 			value: 0n,
-			data: "0x01b443335afe000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000005afe5afe5afe",
+			data: "0x53a8081a5afe000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000005afe5afe5afe",
 			gas: 300_000n,
 		},
 	],
