+++ validator/src/consensus/protocol/onchain.ts
@@ -1,24 +1,18 @@
-import {
-	type Address,
-	type Chain,
-	encodeFunctionData,
-	type FeeValuesEIP1559,
-	type Hex,
-	keccak256,
-	NonceTooLowError,
-	type PublicClient,
-	type SimulateContractParameters,
-	TransactionExecutionError,
-	type Transport,
-} from "viem";
+import { type Address, type Chain, encodeFunctionData, type Hex, type PublicClient, type Transport } from "viem";
 import type { FrostPoint, GroupId, SignatureId } from "../../frost/types.js";
 import { CONSENSUS_FUNCTIONS, COORDINATOR_FUNCTIONS } from "../../types/abis.js";
 import type { ValidatorAccount } from "../../types/account.js";
-import { formatError } from "../../utils/errors.js";
 import type { Logger } from "../../utils/logging.js";
-import { maxBigInt } from "../../utils/math.js";
 import type { Queue } from "../../utils/queue.js";
 import { BaseProtocol, type SubmittedAction } from "./base.js";
+import {
+	type EthTransactionData,
+	type EthTransactionDetails,
+	type FeeValues,
+	GasFeeEstimator,
+	TransactionManager,
+	type TransactionStorage,
+} from "./transaction.js";
 import type {
 	ActionWithTimeout,
 	AttestTransaction,
@@ -35,57 +29,14 @@ import type {
 	StartKeyGen,
 } from "./types.js";
 
-export type FeeValues = Pick<FeeValuesEIP1559, "maxFeePerGas" | "maxPriorityFeePerGas">;
-export type EthTransactionData = { to: Address; value: bigint; data: Hex; gas?: bigint };
-export type EthTransactionDetails = { nonce: number; fees: FeeValues | null; hash: Hex | null };
-
-export interface TransactionStorage {
-	register(tx: EthTransactionData, minNonce: number): number;
-	countPending(): number;
-	delete(nonce: number): void;
-	setPending(nonce: number): void;
-	setFees(nonce: number, fees: FeeValues): void;
-	setHash(nonce: number, txHash: Hex): void;
-	setExecutedUpTo(nonce: number): number;
-	setSubmittedForPending(blockNumber: bigint): number;
-	maxNonce(): number | null;
-	submittedUpTo(blockNumber: bigint, offset?: number, limit?: number): (EthTransactionData & EthTransactionDetails)[];
-}
-
-export class GasFeeEstimator {
-	#cachedPrices: Promise<FeeValues> | null = null;
-	#client: PublicClient;
-
-	constructor(client: PublicClient) {
-		this.#client = client;
-	}
-
-	invalidate() {
-		this.#cachedPrices = null;
-	}
-
-	estimateFees(): Promise<FeeValues> {
-		if (this.#cachedPrices !== null) {
-			return this.#cachedPrices;
-		}
-		// Also cache errors, to prevent that on error too many request are fired
-		const pricePromise = this.#client.estimateFeesPerGas();
-		this.#cachedPrices = pricePromise;
-		return pricePromise;
-	}
-}
+export type { EthTransactionData, EthTransactionDetails, FeeValues, TransactionStorage };
+export { GasFeeEstimator };
 
 export class OnchainProtocol extends BaseProtocol {
 	#publicClient: PublicClient<Transport, Chain>;
-	#account: ValidatorAccount;
-	#gasFeeEstimator: GasFeeEstimator;
-	#txStorage: TransactionStorage;
 	#consensus: Address;
 	#coordinator: Address;
-	#logger: Logger;
-	#blocksBeforeResubmit: bigint;
-	#queuedPendingCheckBlockNumber: bigint | null = null;
-	#runningPendingCheck = false;
+	#txManager: TransactionManager;
 
 	constructor({
 		publicClient,
@@ -110,13 +61,16 @@ export class OnchainProtocol extends BaseProtocol {
 	}) {
 		super(queue, logger);
 		this.#publicClient = publicClient;
-		this.#account = account;
-		this.#gasFeeEstimator = gasFeeEstimator;
-		this.#txStorage = txStorage;
 		this.#consensus = consensus;
 		this.#coordinator = coordinator;
-		this.#logger = logger;
-		this.#blocksBeforeResubmit = blocksBeforeResubmit ?? 1n;
+		this.#txManager = new TransactionManager({
+			publicClient,
+			account,
+			gasFeeEstimator,
+			txStorage,
+			logger,
+			blocksBeforeResubmit,
+		});
 	}
 
 	chainId(): bigint {
@@ -133,151 +87,11 @@ export class OnchainProtocol extends BaseProtocol {
 	}
 
 	isRunningPendingCheck(): boolean {
-		return this.#runningPendingCheck;
+		return this.#txManager.isRunningPendingCheck();
 	}
 
 	triggerPendingCheck(blockNumber: bigint) {
-		if (this.#runningPendingCheck) {
-			this.#logger.debug(`Queueing pending actions check for block ${blockNumber}`);
-			this.#queuedPendingCheckBlockNumber = blockNumber;
-			return;
-		}
-		this.#runningPendingCheck = true;
-		this.#checkPendingActionsLoop(blockNumber).finally(() => {
-			this.#runningPendingCheck = false;
-		});
-	}
-
-	async #checkPendingActionsLoop(initialBlockNumber: bigint) {
-		let blockForCheck: bigint | null = initialBlockNumber;
-		while (blockForCheck !== null) {
-			try {
-				await this.#checkPendingActions(blockForCheck);
-			} catch (e) {
-				this.#logger.error("Error while checking pending transactions.", { error: formatError(e) });
-			}
-			blockForCheck = this.#queuedPendingCheckBlockNumber;
-			this.#queuedPendingCheckBlockNumber = null;
-		}
-	}
-
-	async #checkPendingActions(blockNumber: bigint) {
-		// Optimistically check whether or not we have pending actions. If we don't then we can just exit early and
-		// save on some RPC calls and database reads.
-		if (this.#txStorage.countPending() === 0) {
-			return;
-		}
-
-		// For transaction without a submission block set it to this block
-		// This assumes that the transaction should be included in this block
-		// If the blocksBeforeResubmit is 1 block, these transactions will only be retried on the next block
-		const newPendingTxs = this.#txStorage.setSubmittedForPending(blockNumber);
-		if (newPendingTxs > 0) {
-			this.#logger.debug(`Marked ${newPendingTxs} transactions as submitted at block ${blockNumber}`);
-		}
-		const currentNonce = await this.#publicClient.getTransactionCount({
-			address: this.#account.address,
-			blockTag: "latest",
-		});
-		const executedNonce = currentNonce - 1;
-		const executedTxs = this.#txStorage.setExecutedUpTo(executedNonce);
-		if (executedTxs > 0) {
-			this.#logger.debug(
-				`Marked ${executedTxs} transactions as executed up to nonce ${executedNonce} on block ${blockNumber}`,
-			);
-		}
-		// Only fetch the first page of pending transactions (default limit is 100) to avoid retrying too many
-		// transactions at once.
-		const pendingTxs = this.#txStorage.submittedUpTo(blockNumber - this.#blocksBeforeResubmit);
-		for (const tx of pendingTxs) {
-			this.#logger.debug(`Resubmit transaction for ${tx.nonce}!`, { transaction: tx });
-			try {
-				await this.submitTransaction(tx);
-			} catch (error) {
-				if (
-					error instanceof NonceTooLowError ||
-					// Nonce error might be nested as cause error
-					(error instanceof TransactionExecutionError && error.cause instanceof NonceTooLowError)
-				) {
-					this.#logger.info(`Nonce already used. Marking transaction with nonce ${tx.nonce} as executed!`, {
-						transaction: tx,
-					});
-					this.#txStorage.setExecutedUpTo(tx.nonce);
-					continue;
-				}
-				this.#logger.warn(`Error submitting transaction for ${tx.nonce}!`, { error: formatError(error) });
-				// If an error occurs skip rest of pending transactions, to avoid triggering more errors
-				return;
-			}
-		}
-	}
-
-	private async submitTransaction(
-		tx: EthTransactionData & Pick<EthTransactionDetails, "nonce" | "fees">,
-	): Promise<Hex> {
-		const estimatedFees = await this.#gasFeeEstimator.estimateFees();
-		// Use max of (previous fees + 10%) and estimate
-		const fees: FeeValues = {
-			maxFeePerGas: maxBigInt(estimatedFees.maxFeePerGas, ((tx.fees?.maxFeePerGas ?? 0n) * 110n) / 100n),
-			maxPriorityFeePerGas: maxBigInt(
-				estimatedFees.maxPriorityFeePerGas,
-				((tx.fees?.maxPriorityFeePerGas ?? 0n) * 110n) / 100n,
-			),
-		};
-
-		this.#txStorage.setPending(tx.nonce);
-		// Store fees before submission in case an error occurs
-		this.#txStorage.setFees(tx.nonce, fees);
-		const unsignedTx = {
-			to: tx.to,
-			value: tx.value,
-			data: tx.data,
-			nonce: tx.nonce,
-			gas: tx.gas,
-			chainId: this.#publicClient.chain.id,
-			...fees,
-		};
-		const signedTx = await this.#account.signTransaction(unsignedTx);
-		const txHash = keccak256(signedTx);
-		this.#txStorage.setHash(tx.nonce, txHash);
-		this.#logger.debug(`Submitting transaction for nonce ${tx.nonce}!`, { tx: { hash: txHash, ...unsignedTx } });
-		return this.#publicClient.sendRawTransaction({ serializedTransaction: signedTx });
-	}
-
-	private async submitAction(action: SimulateContractParameters): Promise<SubmittedAction> {
-		// 1. Get Network Baseline (The "Minimum Nonce")
-		// Use 'latest' as all pending transactions are tracked in the tx storage
-		const onChainNonce = await this.#publicClient.getTransactionCount({
-			address: this.#account.address,
-			blockTag: "latest",
-		});
-
-		const calldata = encodeFunctionData(action);
-		// 2. Reserve Nonce & Persist Intent (Atomic DB Operation)
-		// This calculates the correct nonce and saves the record as 'QUEUED'
-		const txData = {
-			to: action.address,
-			data: calldata,
-			value: 0n,
-			gas: action.gas,
-		};
-		const nonce = this.#txStorage.register(txData, onChainNonce);
-		try {
-			const hash = await this.submitTransaction({ ...txData, nonce, fees: null });
-			return { nonce, hash };
-		} catch (err) {
-			// Check if this tx is still the latest, if so delete it and throw error
-			// Retrying should happen on action level, which allows timeouts to apply
-			if (nonce === this.#txStorage.maxNonce()) {
-				this.#txStorage.delete(nonce);
-				throw err;
-			}
-
-			// If another action was already submitted it is important to prevent potential unused nonces
-			// In this case the transaction is kept and retried until executed (to use up the nonce)
-			// No error is thrown to avoid that the action is retried.
-			return { nonce, hash: null };
-		}
+		this.#txManager.triggerPendingCheck(blockNumber);
 	}
 
 	protected startKeyGen({
@@ -290,23 +104,21 @@ export class OnchainProtocol extends BaseProtocol {
 		pok,
 		poap,
 	}: StartKeyGen): Promise<SubmittedAction> {
-		return this.submitAction({
-			address: this.#coordinator,
-			abi: COORDINATOR_FUNCTIONS,
-			functionName: "keyGenAndCommit",
-			args: [
-				participants,
-				count,
-				threshold,
-				context,
-				poap,
-				{
-					q: encryptionPublicKey,
-					c: commitments,
-					r: pok.r,
-					mu: pok.mu,
-				},
-			],
+		return this.#txManager.submitAction({
+			to: this.#coordinator,
+			data: encodeFunctionData({
+				abi: COORDINATOR_FUNCTIONS,
+				functionName: "keyGenAndCommit",
+				args: [
+					participants,
+					count,
+					threshold,
+					context,
+					poap,
+					{ q: encryptionPublicKey, c: commitments, r: pok.r, mu: pok.mu },
+				],
+			}),
+			value: 0n,
 			gas: 250_000n,
 		});
 	}
@@ -316,53 +128,53 @@ export class OnchainProtocol extends BaseProtocol {
 		verificationShare,
 		shares,
 	}: PublishSecretShares): Promise<SubmittedAction> {
-		return this.submitAction({
-			address: this.#coordinator,
-			abi: COORDINATOR_FUNCTIONS,
-			functionName: "keyGenSecretShare",
-			args: [
-				groupId,
-				{
-					y: verificationShare,
-					f: shares,
-				},
-			],
-			gas: 250_000n + BigInt(shares.length) * 25_000n, // TODO: the gas amount per share has not been estimates
+		return this.#txManager.submitAction({
+			to: this.#coordinator,
+			data: encodeFunctionData({
+				abi: COORDINATOR_FUNCTIONS,
+				functionName: "keyGenSecretShare",
+				args: [groupId, { y: verificationShare, f: shares }],
+			}),
+			value: 0n,
+			gas: 250_000n + BigInt(shares.length) * 25_000n, // TODO: the gas amount per share has not been estimated
 		});
 	}
 
 	private confirmKeyGenWithCallback(groupId: GroupId, callbackContext: Hex): Promise<SubmittedAction> {
-		return this.submitAction({
-			address: this.#coordinator,
-			abi: COORDINATOR_FUNCTIONS,
-			functionName: "keyGenConfirmWithCallback",
-			args: [
-				groupId,
-				{
-					target: this.#consensus,
-					context: callbackContext,
-				},
-			],
+		return this.#txManager.submitAction({
+			to: this.#coordinator,
+			data: encodeFunctionData({
+				abi: COORDINATOR_FUNCTIONS,
+				functionName: "keyGenConfirmWithCallback",
+				args: [groupId, { target: this.#consensus, context: callbackContext }],
+			}),
+			value: 0n,
 			gas: 300_000n,
 		});
 	}
 
 	protected complain({ groupId, accused }: Complain): Promise<SubmittedAction> {
-		return this.submitAction({
-			address: this.#coordinator,
-			abi: COORDINATOR_FUNCTIONS,
-			functionName: "keyGenComplain",
-			args: [groupId, accused],
+		return this.#txManager.submitAction({
+			to: this.#coordinator,
+			data: encodeFunctionData({
+				abi: COORDINATOR_FUNCTIONS,
+				functionName: "keyGenComplain",
+				args: [groupId, accused],
+			}),
+			value: 0n,
 			gas: 300_000n, // TODO: this has not been estimated yet
 		});
 	}
 
 	protected complaintResponse({ groupId, plaintiff, secretShare }: ComplaintResponse): Promise<SubmittedAction> {
-		return this.submitAction({
-			address: this.#coordinator,
-			abi: COORDINATOR_FUNCTIONS,
-			functionName: "keyGenComplaintResponse",
-			args: [groupId, plaintiff, secretShare],
+		return this.#txManager.submitAction({
+			to: this.#coordinator,
+			data: encodeFunctionData({
+				abi: COORDINATOR_FUNCTIONS,
+				functionName: "keyGenComplaintResponse",
+				args: [groupId, plaintiff, secretShare],
+			}),
+			value: 0n,
 			gas: 300_000n, // TODO: this has not been estimated yet
 		});
 	}
@@ -371,21 +183,27 @@ export class OnchainProtocol extends BaseProtocol {
 		if (callbackContext !== undefined) {
 			return this.confirmKeyGenWithCallback(groupId, callbackContext);
 		}
-		return this.submitAction({
-			address: this.#coordinator,
-			abi: COORDINATOR_FUNCTIONS,
-			functionName: "keyGenConfirm",
-			args: [groupId],
+		return this.#txManager.submitAction({
+			to: this.#coordinator,
+			data: encodeFunctionData({
+				abi: COORDINATOR_FUNCTIONS,
+				functionName: "keyGenConfirm",
+				args: [groupId],
+			}),
+			value: 0n,
 			gas: 200_000n,
 		});
 	}
 
 	protected requestSignature({ groupId, message }: RequestSignature): Promise<SubmittedAction> {
-		return this.submitAction({
-			address: this.#coordinator,
-			abi: COORDINATOR_FUNCTIONS,
-			functionName: "sign",
-			args: [groupId, message],
+		return this.#txManager.submitAction({
+			to: this.#coordinator,
+			data: encodeFunctionData({
+				abi: COORDINATOR_FUNCTIONS,
+				functionName: "sign",
+				args: [groupId, message],
+			}),
+			value: 0n,
 			gas: 400_000n, // TODO: this has not been estimated yet
 		});
 	}
@@ -394,11 +212,14 @@ export class OnchainProtocol extends BaseProtocol {
 		groupId,
 		nonceCommitmentsHash,
 	}: RegisterNonceCommitments): Promise<SubmittedAction> {
-		return this.submitAction({
-			address: this.#coordinator,
-			abi: COORDINATOR_FUNCTIONS,
-			functionName: "preprocess",
-			args: [groupId, nonceCommitmentsHash],
+		return this.#txManager.submitAction({
+			to: this.#coordinator,
+			data: encodeFunctionData({
+				abi: COORDINATOR_FUNCTIONS,
+				functionName: "preprocess",
+				args: [groupId, nonceCommitmentsHash],
+			}),
+			value: 0n,
 			gas: 250_000n,
 		});
 	}
@@ -408,18 +229,18 @@ export class OnchainProtocol extends BaseProtocol {
 		nonceCommitments,
 		nonceProof,
 	}: RevealNonceCommitments): Promise<SubmittedAction> {
-		return this.submitAction({
-			address: this.#coordinator,
-			abi: COORDINATOR_FUNCTIONS,
-			functionName: "signRevealNonces",
-			args: [
-				signatureId,
-				{
-					d: nonceCommitments.hidingNonceCommitment,
-					e: nonceCommitments.bindingNonceCommitment,
-				},
-				nonceProof,
-			],
+		return this.#txManager.submitAction({
+			to: this.#coordinator,
+			data: encodeFunctionData({
+				abi: COORDINATOR_FUNCTIONS,
+				functionName: "signRevealNonces",
+				args: [
+					signatureId,
+					{ d: nonceCommitments.hidingNonceCommitment, e: nonceCommitments.bindingNonceCommitment },
+					nonceProof,
+				],
+			}),
+			value: 0n,
 			gas: 200_000n,
 		});
 	}
@@ -434,27 +255,20 @@ export class OnchainProtocol extends BaseProtocol {
 		lagrangeCoefficient: bigint,
 		callbackContext: Hex,
 	): Promise<SubmittedAction> {
-		return this.submitAction({
-			address: this.#coordinator,
-			abi: COORDINATOR_FUNCTIONS,
-			functionName: "signShareWithCallback",
-			args: [
-				signatureId,
-				{
-					r: groupCommitment,
-					root: signersRoot,
-				},
-				{
-					r: commitmentShare,
-					z: signatureShare,
-					l: lagrangeCoefficient,
-				},
-				signersProof,
-				{
-					target: this.#consensus,
-					context: callbackContext,
-				},
-			],
+		return this.#txManager.submitAction({
+			to: this.#coordinator,
+			data: encodeFunctionData({
+				abi: COORDINATOR_FUNCTIONS,
+				functionName: "signShareWithCallback",
+				args: [
+					signatureId,
+					{ r: groupCommitment, root: signersRoot },
+					{ r: commitmentShare, z: signatureShare, l: lagrangeCoefficient },
+					signersProof,
+					{ target: this.#consensus, context: callbackContext },
+				],
+			}),
+			value: 0n,
 			gas: 400_000n,
 		});
 	}
@@ -481,23 +295,19 @@ export class OnchainProtocol extends BaseProtocol {
 				callbackContext,
 			);
 		}
-		return this.submitAction({
-			address: this.#coordinator,
-			abi: COORDINATOR_FUNCTIONS,
-			functionName: "signShare",
-			args: [
-				signatureId,
-				{
-					r: groupCommitment,
-					root: signersRoot,
-				},
-				{
-					r: commitmentShare,
-					z: signatureShare,
-					l: lagrangeCoefficient,
-				},
-				signersProof,
-			],
+		return this.#txManager.submitAction({
+			to: this.#coordinator,
+			data: encodeFunctionData({
+				abi: COORDINATOR_FUNCTIONS,
+				functionName: "signShare",
+				args: [
+					signatureId,
+					{ r: groupCommitment, root: signersRoot },
+					{ r: commitmentShare, z: signatureShare, l: lagrangeCoefficient },
+					signersProof,
+				],
+			}),
+			value: 0n,
 			gas: 400_000n,
 		});
 	}
@@ -509,31 +319,40 @@ export class OnchainProtocol extends BaseProtocol {
 		safeTxStructHash,
 		signatureId,
 	}: AttestTransaction): Promise<SubmittedAction> {
-		return this.submitAction({
-			address: this.#consensus,
-			abi: CONSENSUS_FUNCTIONS,
-			functionName: "attestTransaction",
-			args: [epoch, chainId, safe, safeTxStructHash, signatureId],
+		return this.#txManager.submitAction({
+			to: this.#consensus,
+			data: encodeFunctionData({
+				abi: CONSENSUS_FUNCTIONS,
+				functionName: "attestTransaction",
+				args: [epoch, chainId, safe, safeTxStructHash, signatureId],
+			}),
+			value: 0n,
 			gas: 400_000n, // TODO: this has not been estimated yet
 		});
 	}
 
 	protected stageEpoch({ proposedEpoch, rolloverBlock, groupId, signatureId }: StageEpoch): Promise<SubmittedAction> {
-		return this.submitAction({
-			address: this.#consensus,
-			abi: CONSENSUS_FUNCTIONS,
-			functionName: "stageEpoch",
-			args: [proposedEpoch, rolloverBlock, groupId, signatureId],
+		return this.#txManager.submitAction({
+			to: this.#consensus,
+			data: encodeFunctionData({
+				abi: CONSENSUS_FUNCTIONS,
+				functionName: "stageEpoch",
+				args: [proposedEpoch, rolloverBlock, groupId, signatureId],
+			}),
+			value: 0n,
 			gas: 400_000n, // TODO: this has not been estimated yet
 		});
 	}
 
 	protected setValidatorStaker({ staker }: SetValidatorStaker): Promise<SubmittedAction> {
-		return this.submitAction({
-			address: this.#consensus,
-			abi: CONSENSUS_FUNCTIONS,
-			functionName: "setValidatorStaker",
-			args: [staker],
+		return this.#txManager.submitAction({
+			to: this.#consensus,
+			data: encodeFunctionData({
+				abi: CONSENSUS_FUNCTIONS,
+				functionName: "setValidatorStaker",
+				args: [staker],
+			}),
+			value: 0n,
 			gas: 400_000n, // TODO: this has not been estimated yet
 		});
 	}
